+++ zia/common/resourcetype/resource_type.go
@@ -32,4 +32,5 @@ const (
 	SSLInspectionRules              = "zia_ssl_inspection_rules"
 	AuthSettingsURLs                = "zia_auth_settings_urls"
 	Activation                      = "zia_activation_status"
+	AdminRoles                      = "zia_admin_roles"
 )